Improved logInMenu() in Main.
The previous implementation of the function made it so that if a user were to
press login instead of register by accident, he would get stuck with a login
prompt and would be forced to use Ctrl+C to exit. Changes were made so that this
doesn't happen.

I also implemented the exit functionality and made sure that a message was shown
for invalid inputs.